Backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-stubs==1.10.1
django-stubs-ext==0.4.0
filelock==3.6.0
gprof2dot==2021.2.21
graphene==3.0
graphene-django==3.0.0b7
graphene-django-optimizer==0.9.1
graphql-core==3.1.2
graphql-relay==3.1.5
gunicorn==20.1.0
hiredis==2.0.0
identify==2.4.11
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
mypy==0.942
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pre-commit==2.17.0
promise==2.3
psycopg2==2.9.3
pycodestyle==2.8.0
pyinstrument==4.1.1
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
types-pytz==2021.3.6
types-PyYAML==6.0.5
typing-extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.3
wrapt==1.14.0
